+++ ST/Smolnyakov Anton/project_max/src/components/index.js
@@ -4,7 +4,6 @@ import header from "./js/header"
 export default () => {
     let cart = new Cart('/cart.json', '#cart');
     let catalog = new Catalog('/catalog.json', '#catalog', cart);
-    console.log(cart, catalog)
     header.handleEvents();
 }
