+++ ST/Smolnyakov Anton/project_max/src/components/js/parents.js
@@ -1,10 +1,10 @@
-export default class parentCartAndCatalog {
-    constructor(cart) {
-        this.items = [];
-        this.container = null;
-        this.cart = cart;
-        this.totalSumContainer = null;
-        this.totalSum = 0;
-        this._init()
-    }
-}
+// export default class parentCartAndCatalog {
+//     constructor(cart) {
+//         this.items = [];
+//         this.container = null;
+//         this.cart = cart;
+//         this.totalSumContainer = null;
+//         this.totalSum = 0;
+//         this._init()
+//     }
+// }
